# ansi: speed up escape sequence parsing by ~20%