This commit slightly refactors the escape sequence matching functions
for a 20% speed up.

ANSI Benchmarks:
```
goos: darwin
goarch: arm64
pkg: github.com/junegunn/fzf/src

name                       old time/op    new time/op    delta
NextAnsiEscapeSequence-10     114ns ± 1%      90ns ± 1%  -21.57%  (p=0.008 n=5+5)
ExtractColor-10               544ns ± 0%     517ns ± 0%   -5.01%  (p=0.008 n=5+5)

name                       old speed      new speed      delta
NextAnsiEscapeSequence-10  1.09GB/s ± 1%  1.39GB/s ± 1%  +27.52%  (p=0.008 n=5+5)
ExtractColor-10             230MB/s ± 0%   242MB/s ± 0%   +5.27%  (p=0.008 n=5+5)
```