@junegunn 2022-08-18
Can you share the commands you used to get the result you posted above? Here's what I'm seeing, a bit different from yours.

```sh
go install golang.org/x/perf/cmd/benchstat
make bench > old.txt
hub pr checkout 2927
make bench > new.txt
$GOPATH/bin/benchstat old.txt new.txt
```

```
name                             old time/op    new time/op    delta
NextAnsiEscapeSequence-12           140ns ± 0%     134ns ± 0%   ~     (p=1.000 n=1+1)
NextAnsiEscapeSequence_Regex-12    4.30µs ± 0%    4.13µs ± 0%   ~     (p=1.000 n=1+1)
ExtractColor-12                     783ns ± 0%     737ns ± 0%   ~     (p=1.000 n=1+1)

name                             old speed      new speed      delta
NextAnsiEscapeSequence-12         893MB/s ± 0%   934MB/s ± 0%   ~     (p=1.000 n=1+1)
NextAnsiEscapeSequence_Regex-12  29.1MB/s ± 0%  30.3MB/s ± 0%   ~     (p=1.000 n=1+1)
ExtractColor-12                   160MB/s ± 0%   170MB/s ± 0%   ~     (p=1.000 n=1+1)

name                             old alloc/op   new alloc/op   delta
NextAnsiEscapeSequence-12           0.00B          0.00B        ~     (all equal)
NextAnsiEscapeSequence_Regex-12      224B ± 0%      224B ± 0%   ~     (all equal)
ExtractColor-12                      312B ± 0%      312B ± 0%   ~     (all equal)

name                             old allocs/op  new allocs/op  delta
NextAnsiEscapeSequence-12            0.00           0.00        ~     (all equal)
NextAnsiEscapeSequence_Regex-12      14.0 ± 0%      14.0 ± 0%   ~     (all equal)
ExtractColor-12                      4.00 ± 0%      4.00 ± 0%   ~     (all equal)
```