@charlievieth 2022-08-25
> Can you share the commands you used to get the result you posted above? Here's what I'm seeing, a bit different from yours.

Thanks for double checking! I originally ran the benchmarks on a M1 Mac so I re-ran them on my Intel i9-9900K machine and get results more similar to yours (though still seeing a ~6% speedup):

```
goos: windows
goarch: amd64
pkg: github.com/junegunn/fzf/src
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz

name                       old time/op   new time/op    delta
NextAnsiEscapeSequence-16    130ns ± 3%     122ns ± 2%  -6.35%  (p=0.008 n=5+5)
ExtractColor-16              720ns ± 1%     687ns ± 1%  -4.52%  (p=0.008 n=5+5)

name                       old speed     new speed      delta
NextAnsiEscapeSequence-16  959MB/s ± 3%  1024MB/s ± 2%  +6.76%  (p=0.008 n=5+5)
ExtractColor-16            174MB/s ± 1%   182MB/s ± 1%  +4.73%  (p=0.008 n=5+5)
```

For reference, I'm using the following for benchmarks: 
```sh
$ git checkout master
$ go test -run '^$' -bench '(NextAnsiEscapeSequence| ExtractColor)$' -count 5 | tee old.txt
$ git checkout cev/ansi-perf
$ go test -run '^$' -bench '(NextAnsiEscapeSequence| ExtractColor)$' -count 5 | tee new.txt
$ benchstat old.txt new.txt
```