@junegunn 2022-08-25
Merged, thanks!